[chance0523-issue2] 코드 리뷰 반영
