+++ src/main/java/com/study/realworld/user/application/UserService.java
@@ -2,11 +2,9 @@ package com.study.realworld.user.application;
 
 import com.study.realworld.core.domain.user.entity.User;
 import com.study.realworld.core.domain.user.repository.UserRepository;
-import com.study.realworld.user.presentation.model.UserRegisterModel;
+import com.study.realworld.user.application.model.UserRegisterModel;
 
 import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
 
 import javax.transaction.Transactional;
 
@@ -23,12 +21,6 @@ public class UserService {
 
     @Transactional
     public User register(UserRegisterModel userRegisterModel) {
-        User user = User.builder()
-                        .userName(userRegisterModel.getUserName())
-                        .email(userRegisterModel.getEmail())
-                        .password(userRegisterModel.getPassword())
-                        .createdAt(LocalDateTime.now())
-                        .build();
-        return userRepository.save(user);
+        return userRepository.save(userRegisterModel.toUser());
     }
 }
